Lots of stuff: Ports from kwcoco, download manager, candidate acceptance.
